change int to size_t

+++ cpp/MTB/mtb.hpp
@@ -6,7 +6,7 @@
 template <typename T>
 std::ostream& operator<<(std::ostream& os, const std::vector<T>& vec) {
     os << "[";
-    for (int i = 0; i < vec.size(); ++i) {
+    for (size_t i = 0; i < vec.size(); ++i) {
         os << vec[i];
         if (i != vec.size() - 1)
             os << ", ";
