+++ cpp/MTB/test/test_common.hpp
@@ -6,7 +6,7 @@
 template <typename T>
 bool compare_vectors(const std::vector<T>& a, const std::vector<T>& b) {
     if (a.size() != b.size()) return false;
-    for (int i = 0; i < static_cast<int>(a.size()); ++i) {
+    for (size_t i = 0; i < static_cast<size_t>(a.size()); ++i) {
         if (a[i] != b[i]) return false;
     }
     return true;
